Documented getConcept() method, added ontology classes
The API should ideally return all the parent classes for a concept, but until then, the library does a lookup.

It is using a local copy of the dbpedia 3.9 ontology extracted by parsing http://mappings.dbpedia.org/server/ontology/classes/

Note: Currently the API doesn't return the type information unless you do getConcepts() or getConcept() - explicit information is not returned when viewing tags on articles, for example.

I'll see if I can get this as an enhancement in the API before I try and fix it here (as having the API do it would be much faster).