Improving input handling and updating package data
Using parseInt() on input - and default to zero if the parse fails - so that methods cope gracefully with numerical values that are passed as strings.

This makes consuming the library easier and reduces the amount of special case handling needed client side when building applications that consume the library.

I've also updated the package metadata to reflect that the library is now in the BBC News Labs repo.